Create python package for sensor commands example
