+++ examples/python/sensor_commands/README.md
@@ -1,0 +1,3 @@
+# Sensor commands package
+
+This is an example package for the sensor commands modules.
