+++ examples/python/sensor_commands/config/sensors_cfg.json
