+++ examples/python/sensor_commands/lib/sensor_commands/__init__.py
@@ -1,0 +1,7 @@
+"""
+sensor_commands module entry point.
+"""
+
+__author__ = 'Esteban Zamora Alvarado'
+__email__ = 'esteban.zamora.al@gmail.com'
+__version__ = '0.0.1'
