+++ examples/python/sensor_commands/lib/sensor_commands/main.py
@@ -1,17 +1,34 @@
-import command
-import sensor
-from random import choice
+import logging
 from time import sleep
+from random import choice
+
+from . import command
+from .args import parse_args
+from .sensor.manager import SensorManager
+from .sensor.analyzers import avg_thresh_analyzer as avt
+
+
+log = logging.getLogger(__name__)
+
 
 def main():
-    config_name = "sensor_config.json"
-    sensor_type_name = "temperature"
+    """
+    Application main function.
+    :param args: An arguments namespace.
+    :type args: :py:class:`argparse.Namespace`
+    :return: Exit code.
+    :rtype: int
+    """
+    args = parse_args()
+
+    config_name = args.config
+    sensor_type_name = args.sensor_type
     sensor_cmd_per_period, sensor_period_sec = (100, 5)
     alert_cmd_per_period, alert_period_sec = (2, 1)
     analyzer_avg_thresh = 10
     num_read_commands = 200
 
-    sensor_mgr = sensor.SensorManager(config_name)
+    sensor_mgr = SensorManager(config_name)
     sensor_cmd_runner = command.CommandRunner(cmd_per_period=sensor_cmd_per_period,
                                               period_sec=sensor_period_sec)
     alert_cmd_runner = command.CommandRunner(cmd_per_period=alert_cmd_per_period,
@@ -20,9 +37,9 @@ def main():
     alert_cmd_runner.start()
 
     # Set up sensor analyzer with "above average threshold alert" strategy
-    analyzer = sensor.SensorAvgThreshAnalyzer(avg_thresh=analyzer_avg_thresh)
-    sensor.set_avg_thresh_alert_handle_strategy(analyzer, alert_cmd_runner)
-    sensor.set_above_avg_thresh_compare_strategy(analyzer)
+    analyzer = avt.SensorAvgThreshAnalyzer(avg_thresh=analyzer_avg_thresh)
+    avt.set_alert_handle_strategy(analyzer, alert_cmd_runner)
+    avt.set_above_compare_strategy(analyzer)
 
     # Generate read commands for temp sensors
     temp_sensor_names = sensor_mgr.get_sensor_names_per_type(sensor_type_name)
@@ -33,6 +50,7 @@ def main():
 
     sensor_cmd_runner.stop()
     alert_cmd_runner.stop()
+
 
 if __name__ == "__main__":
     main()
