+++ examples/python/sensor_commands/lib/sensor_commands/sensor/__init__.py
@@ -1,0 +1,3 @@
+"""
+sensor module entry point.
+"""
