+++ examples/python/sensor_commands/lib/sensor_commands/sensor/analyzers/__init__.py
@@ -1,0 +1,3 @@
+"""
+sensor analyzers module entry point.
+"""
