+++ examples/python/sensor_commands/lib/sensor_commands/sensor/sensors/__init__.py
@@ -1,0 +1,4 @@
+"""
+sensors module entry point.
+"""
+from .factory import SensorFactory
