+++ examples/python/sensor_commands/lib/sensor_commands/sensor/sensors/level.py
@@ -1,0 +1,11 @@
+from random import randint
+
+from ..sensor import Sensor
+
+
+class LevelSensor(Sensor):
+    def __init__(self, name: str) -> None:
+        super().__init__(name, "level", "meters")
+
+    def read(self) -> float:
+        return randint(1, 50)
