+++ examples/python/sensor_commands/requirements.txt
@@ -1,0 +1,1 @@
+ratelimiter==1.2.0.post0
